crm1/requirements.txt
asgiref==3.7.2
boto3==1.34.44
botocore==1.34.44
Django==5.0.1
django-filter==23.5
django-session-timeout==0.1.0
django-storages==1.14.2
greenlet==3.0.3
jmespath==1.0.1
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
psycopg==3.1.18
psycopg2==2.9.9
pyarrow==15.0.0
python-dateutil==2.8.2
pytz==2023.4
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7

